+++ backend/service-lib/src/main/kotlin/calls/ApiConventions.kt
@@ -21,7 +21,7 @@ object UCloudApi {
 /**
  * Used for RPCs which request the creation of one or more resources in UCloud.
  *
- * RPCs in this category must accept request payloads in the form of a
+ * RPCs in this category should accept request payloads in the form of a
  * [bulk request](/backend/service-lib/wiki/api_conventions.md#bulk-request).
  *
  * Calls in this category should respond back with a list of newly created IDs for every resource that has been
@@ -377,7 +377,7 @@ private fun CallDescriptionContainer.httpSearchExample() {
 /**
  * _Interacts_ with a UCloud resource, typically causing an update to the resource itself.
  *
- * RPCs in this category must accept request payloads in the form of a
+ * RPCs in this category should accept request payloads in the form of a
  * [bulk request](/backend/service-lib/wiki/api_conventions.md#bulk-request).
  *
  * This category of calls can be used for any type of interaction with a UCloud resource. Many resources in UCloud
@@ -454,7 +454,7 @@ private fun CallDescriptionContainer.httpUpdateExample() {
 /**
  * Used for RPCs which request the deletion of one or more resources in UCloud.
  *
- * RPCs in this category must accept request payloads in the form of a
+ * RPCs in this category should accept request payloads in the form of a
  * [bulk request](/backend/service-lib/wiki/api_conventions.md#bulk-request).
  *
  * Calls in this category should choose to accept as little information about the resources, while still uniquely
@@ -740,7 +740,7 @@ differences later.
 
 ### Authentication
 
-You can read more about authentication using the different network protocols [here](/backend/auth-service/README.md).
+You can read more about authentication using the different network protocols [here](/docs/developer-guide/core/users/authentication/users.md).
 
 ### Data Serialization
 
@@ -755,13 +755,13 @@ there are significant benefits in supporting multiple representation of the same
 
 UCloud will include a valid `Content-Type` header in response payloads. This content-type will only describe the format
 used to serialize the data. It will not describe the contents of the data. That is, UCloud assumes that clients are
-already aware of the data that any specfic RPC produces.
+already aware of the data that any specific RPC produces.
 
 ---
 
 UCloud attempts to follow the principal of relaxed requirements for request payloads while following a more strict and
 consistent approach for response payloads. You can read more about how we implement this in
-practice [here](/backend/service-lib/wiki/micro/serialization.md).
+practice [here](/docs/developer-guide/development/micro/serialization.md).
 
 ## RPC Interfaces
 
@@ -770,7 +770,7 @@ about or interact with. For example, clients of UCloud can ask about the status 
 UCloud to cancel an existing compute job.
 
 All resources in UCloud have a chunk of the URL path carved out for them. This 'namespace' serves as their area and all
-requests are generally routed in the same direction. For example, [avatars](/backend/avatar-service/README.md) uses
+requests are generally routed in the same direction. For example, [avatars](/docs/developer-guide/core/users/avatars.md) uses
 `/api/avatar` for all their requests. This namespace is commonly referred to as the `baseContext` inside of UCloud.
 
 All RPCs aim to either describe the state of resources in UCloud or interact with one or more resources. RPCs of UCloud
@@ -791,7 +791,7 @@ fall into one of the following categories:
 
 __📝 NOTE:__ Resources don't have to implement one of each call category. For example, some resources cannot be created
 explicitly by a client and can only be interacted with. For example, all UCloud users have exactly one
-[avatar](/backend/avatar-service/README.md). Clients can choose to update their avatar, but they cannot create a new
+[avatar](/docs/developer-guide/core/users/avatars.md). Clients can choose to update their avatar, but they cannot create a new
 avatar in the system.
 
 ---
@@ -992,7 +992,7 @@ private fun CallDescriptionContainer.httpSearchExample() {
 <!--<editor-fold desc="Generated documentation">-->
 Used for RPCs which request the creation of one or more resources in UCloud.
 
-RPCs in this category must accept request payloads in the form of a bulk request.
+RPCs in this category should accept request payloads in the form of a bulk request.
 
 Calls in this category should respond back with a list of newly created IDs for every resource that has been
 created. A client can choose to use these to display information about the newly created resources.
@@ -1044,7 +1044,7 @@ private fun CallDescriptionContainer.httpCreateExample() {
 <!--<editor-fold desc="Generated documentation">-->
 Used for RPCs which request the deletion of one or more resources in UCloud.
 
-RPCs in this category must accept request payloads in the form of a bulk request.
+RPCs in this category should accept request payloads in the form of a bulk request.
 
 Calls in this category should choose to accept as little information about the resources, while still uniquely
 identifying them.
@@ -1091,8 +1091,7 @@ private fun CallDescriptionContainer.httpDeleteExample() {
 <!--<editor-fold desc="Generated documentation">-->
 _Interacts_ with a UCloud resource, typically causing an update to the resource itself.
 
-RPCs in this category must accept request payloads in the form of a
-[bulk request](/backend/service-lib/wiki/api_conventions.md#bulk-request).
+RPCs in this category should accept request payloads in the form of a bulk request.
 
 This category of calls can be used for any type of interaction with a UCloud resource. Many resources in UCloud
 are represented by some underlying complex logic. As a result, it is typically not meaningful or possible to simple
@@ -1491,9 +1490,7 @@ There are several levels of internal APIs:
 
 - __Beta:__
     - Breaking changes can be made quickly (with no real deprecation cycle)
-    - Documentation is optional
 - __Stable:__
-    - Documentation is mandatory
     - Breaking changes should be made with a deprecation cycle lasting two versions
     - Note: The deprecation cycle is only there to allow for rolling upgrades
     - The two versions participating in the deprecation cycle could be rolled out very quickly (less than one day is
@@ -1512,13 +1509,11 @@ There are several levels of experimental APIs:
     - Feature complete according to initial design
     - Breaking changes are made when needed
     - No migration path or deprecation cycle
-    - Documentation is optional
 - __Beta:__
     - Feature complete
     - User feedback is heavily encouraged during this phase
     - Breaking changes are generally only done if feedback suggests that this is needed
     - Short deprecation period if deemed necessary
-    - Documentation is mandatory
 
 External clients are encouraged to use beta-level APIs and to try out alpha-level APIs.