+++ wiki/release-notes.md
@@ -7,9 +7,66 @@ use.
 Note that each service may have more than one deployment associated with it.
 For details look in the `k8.kts` folders of each microservice.
 
+## 15-02-23
+
+```
+Type 1/SDU Provider @ 2022.3.24 (also known as 2022.3.23-im-fix.5)
+```
+
+## 14-02-23
+
+```
+Type 1/SDU Provider @ 2022.3.23-im-fix.3
+```
+
+## 13-02-23
+
+```
+Type 1/SDU Provider @ 2022.3.22
+```
+
+## 10-02-23
+
+```
+Type 1/SDU Provider @ 2022.3.21 (Rolled back to 2022.3.20)
+```
+
+## 09-02-23
+
+```
+✅  Deployment(app-orchestrator, 2022.3.18) (UP-TO-DATE)
+```
+
+## 08-02-23
+
+```
+✅  Deployment(accounting, 2022.3.15) (UP-TO-DATE)
+Type 1/SDU Provider @ 2022.3.20
+```
+
+## 07-02-23
+
+```
+✅  Deployment(webclient, 2022.3.14) (UP-TO-DATE)
+-- app-kubernetes has been turned off --
+-- file-ucloud has been turned off --
+
+Type 1/SDU Provider @ 2022.3.14
+```
+
+## 03-02-23
+
+```
+✅  Deployment(app-store, 2022.3.11) (UP-TO-DATE)
+✅  Deployment(auth, 2022.3.11) (UP-TO-DATE) (rolled back to 2022.2.76)
+```
+
 ## 02-02-23
 
 ```
+✅  Deployment(webclient, 2022.3.10) (UP-TO-DATE)
+✅  Deployment(app-orchestrator, 2022.3.8) (UP-TO-DATE)
+✅  Deployment(file-orchestrator, 2022.3.8) (UP-TO-DATE)
 ✅  Deployment(elastic-management-service, 2022.3.8-elasticfix) (UP-TO-DATE)
 ```